Move to System-as-root
This is required as part of the project Trebble and it also has
several technical advantages over the older solutions like the
ability to use bigger kernel images.

See the following documentation for more details on system as root:
https://source.android.com/devices/bootloader/system-as-root

The cache partition is also not used anymore:
  # ls -ld /cache
  lrw-r--r-- 1 root root 11 2008-12-31 23:00 /cache -> /data/cache
There is more information on that here:
https://source.android.com/devices/tech/ota/ab/ab_faqs

As for the EFS removal: since / is now read-only it's not possible
anymore to create an /efs directory from the init.smdk4x12.rc, so
we remove it for now. Ideally we'd need to find a way, from this
device repository, to tell build/make to create that directory.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>