Add support for flashing through heimdall
The images built are smaller than the partition sizes. This
enables to flash them faster and reduce the probability
of failing the flashing.

The downside is that, for the data partition, the space available
is not enough to complete boot:
  # df -h
  Filesystem            Size  Used Avail Use% Mounted on
  [...]
  /dev/block/mmcblk2p16 6.4M  6.1M  308K  96% /data

The resizing is not done automatically because it's supposed to
happpen during the installation of the zip file.

This makes sures that the data partition is resized during boot.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>