HACK: Prevent the device to go in suspend
This is because it's anoying to have the device power off when
we're using it though adb.

This is for early developement only as it is configurable in
the settings->developer menu and it's default value is here:
framework/base/packages/SettingsProvider/res/values/defaults.xml

It is not desirable to keep that hack in the long run as it
makes the device consume power power, which results in a shorter
battery life in hours.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>