HACK: Add userspace support for i9300
This is to add the least amount of changes to add support
for the i9300.

The changes were first shared with i9305 but that had
several issues:
- We don't support the modem yet on the i9305
- The partitions are different, but the i9300 kernel can be
  built separately by hand, to keep changes minimal.
- Right now we also support the stock bootloader as u-boot
  still requires shipping nonfree software. This requires
  to hardcode the dtb.

To use a single repository, we would need to:
- Have a free software u-boot. If it's not possible to
  replace the nonfree BL1, we could still make an u-boot
  version that can be instealled instead of the kernel
  and use that to handle the different dtbs.
- Add support for the i9305 modem through the ofono ril.
- Add support for libsamsung-ipc in ofono.

Addding a replicant_common repository would also require
more work, so it can be done later.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Rebased for Replicant 10
Signed-off-by: Joonas Kylmälä <joonas.kylmala@iki.fi>