Properly inherit from LineageOS (call to nonexisting file)
Inheriting from vendor/lineage/config/common_phone.mk
does not work because this file does not exist.
Use vendor/lineage/config/common_full.mk instead.

This bug has been caused by the following commit:
    f1555e9 HACK: Add userspace support for i9300

That commit was part of a branch to enable the modem
support for the GT-I9300, and was subsequently fixed in the
following commit in that same branch:
    9aa8650: Add support for the modem

Howver as the fix wasn't merged in f1555e9, the bug was
carried over to replicant 9 and 10 branches.

Signed-off-by: Tobias Tefke <t.tefke@stud.fh-sm.de>
GNUtoo@cyberdimension: Added a clarfication about the bug provenance
Acked-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>