Correct graphics settings
The hardware seems to support vsync, at least no erros come without
this setting. Overlay planes also work, the issue with graphics
corruption was previously because we allocated graphics buffers for
SCANOUT (buffers to be shown on display) on the GPU render node which
seems to allocate the memory non-contiguously with MMU. The exynos
display controller seems to require contiguous memory. When
gralloc.gbm.device is card0 and kmsro driver is enabled then it
handles the allocation between card and render nodes correctly.

Signed-off-by: Joonas Kylmälä <joonas.kylmala@iki.fi>