Add libsamsung-ril
While the Galaxy SIII (GT-I9300) modem isn't supported yet in Replicant 11,
it is still a good idea to start shipping libsamsung-ril as it enables to
test if it compiles fine under Replicant 11.

In addition, as of now, with no modem driver, once booted the CPU
is mostly idle. This has been verified with top which shows
398%idle (there are 4 cores).

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>